refactor apiserver tests
